Updated openstack/openstack Project: openstack/trove-specs  912d650288e9179c93c58d148787dce99d009670
Datastore Registration API spec

Spec for adding the APIs to manage
registration/de-registration of datastore versions
with their image-ids on an existing trove deployment.

In order to simplify the process of registering an image for a new
datastore/datastore-version the datastore would be managed behind the
scenes. This means that when creating a datastore-version, a datastore
will be automatically created, if it does not already exist.

Change-Id: I5631588cf09083987498555b459d3703b6c87d37
Implements: blueprint datastore-registration-api