Updated openstack/openstack Project: openstack/nova  3c6a6332d6f6fc6cc7170003ec0bfcc59017aa24
Don't pass the service catalog when making glance requests

If your service catalog has too many endpoints in it, passing this in
the request to glance can result in a 400 because the header is too big.

This isn't even necessary to pass to glance since we use the auth_token
middleware.

Per bknudson: "Any service using the auth_token middleware has no use
for the x-service-catalog header. All that auth_token middleware uses is
x-auth-token. The auth_token middleware will actually strip the
x-service-catalog from the request before it sends the request on to the
rest of the pipeline, so the application will never see it."

Closes-Bug: #1482699

Change-Id: I204c6f61194bef6eba01ab0525bc80fa2e323acd