Updated openstack/openstack Project: openstack/keystone  1a833c0b93e278dcecd29cdb30d2c7fa03946de9
Fernet 'expires' value loses 'ms' after validation

When we create a Fernet token we get a token with an expiration
time which contains microseconds. When we validate a Fernet token,
the expiration time changes and comes back missing the microseconds.
We need to make the expiration time the same between creation and
validation time, having microseconds in both cases. This applies to
both v2 and v3 API.

Closes-Bug: #1459790
Change-Id: Ic6adbc2d69af03519a2c4ef66d558bcff9022049