Updated openstack/openstack Project: openstack/cinder  993ad796cdff06166962e513d3fc753b07b01eca
Snap copy feature for EMC VNX Cinder driver

Currently in VNX Cinder driver, a full-copy will be performed
when creating a volume from snapshot or volume and it is
time-consuming. This patch is going to address this issue by
providing an alternative way which is performing a snap-copy
when creating a volume from snapshot or volume. This can
accelerate the proccess a lot. And if the user wants to perform
a full-copy on the snap-copy volume, migration or retype can be
triggered on it. And we implement update_migrated_volume to
support migration and retype on the snap-copy volume.

DocImpact
Change-Id: I134a2d2274b1e3b85ea08cbf5e034f47d98e97db
Closes-Bug: #1457274