Updated openstack/openstack Project: openstack/neutron  c28c82e8cdab9668e436f46bf87c34970b54ca58
Make '_create_router' function handle Boolean kwargs correctly

The function allows arbitrary kw arguments to be specificed.
Those kw arguments whose key is specified in the function
parameter 'arg_list' are then passed in the create router REST
API call.

The code line that checks if a kw argument key is included in
'arg_list' does not work with Boolean kw arguments. Nor does it
work with None kw argument values. This is a limitation as
tests in inheriting classes may need such kw arguments.

This patch fixes the faulty 'if' clause so that it simply
checks that the kw argument key is in 'arg_list'. This makes it
support Boolean kw arguments and kw arguments with value None.

Closes-Bug: #1482108
Change-Id: I5618dc4d5c803c7614dd1f579db3c79928007fb6