Updated openstack/openstack Project: openstack/manila  639d0913b98b856662a64707d7aad9ef2ac6923a
Fix 'hacking' unit tests for py3 compatibility

Fix two 'hacking' unit tests, where exist following errors:
- 'str' in py3 is same as 'unicode' in py2, because of this hacking rule test
for violating usage of 'str' fails. Skip it for py3.
- Consider python version in test with translation where we have different
column numbers with errors. In py2 it is starting point of called command and
in py3 it is last point. So, the difference is length of used command that
violates the hacking rule.

Change-Id: Id3d196d6006e7b2e5e56343bd7ba207c5196b61d
Partially-Implements: bp py3-compatibility