Updated openstack/openstack Project: openstack-infra/puppet-httpd  0b6f95fe3f348337c689278afcb207d1cc579566
Fix $httpd::vhost::proxy::servername parameter

The $httpd::vhost::proxy::servername parameter was not being used, and
the local variable $srvname was defaulting to the name of the defined
type. This patch copies the logic from httpd::vhost to conditionally
set $srvname to the $servername parameter if it is set, and default to
the old behavior if not.

Change-Id: I4936be856db1a1af279b7c3824606e2c0375e9a6