Updated openstack/openstack Project: openstack-infra/system-config  97cf312d7fb325913b50f48f874f51dc9fa73b91
Switch ask.openstack.org to Ubuntu Trusty

As part of the ask.openstack.org upgrade, it must be migrated to
Ubuntu Trusty. The askbot puppet module no longer properly supports
Precise and is not being applied on the old ask production server in
preparation for this migration. Go ahead and start testing for the
correct platform now.

Change-Id: I50a991ec9343fae1329601036776bcee7a7a62aa