Updated openstack/openstack Project: openstack/cinder  7584925f4335329c34ee3578f003073c54f47386
Ignore InsecureReq warning in SolidFire Driver

The python requests package logs the use of Insecure as
a warning.  Currently the use of http vs https is a
configurable option, and for those that use http, the
log files include a Warning message for every single
request to the SolidFire Cluster.

In our case we're explicitly providing the
option to use http instead of https, so we can safely
ignore this message and remove it from the logs.

Change-Id: I746b5f7d7351ef594bf0004ec76458fc1a633031