Updated openstack/openstack Project: openstack-infra/project-config  c4e160b05d1e583584fb669581e0e78054a3dc96
Ensure anvil-job has newest six library upgraded

The built-in six library being so old causes the building
of python-cryptography into an rpm to eventually fail at build
time. To address this, make sure that the six installed is
upgraded to be the newest version (which should be safe to do
as six is very much backwards compat.).

Change-Id: I7c7468897f002bb467a78ceb1dcab2d19978c4c1
Closes-Bug: #1482696