Updated openstack/openstack Project: openstack-dev/pbr  5c0bb9186fe2d65901744b00af24c8c50b3e1d29
Handle git being entirely absent

When we try to find the git directory, if git is not present, treat
that the same as an error from git itself. Sadly tests for this are
nearly impossible to meaningfully write, since we always have git
installed. I can do a mock based test if folk want one.

Change-Id: If6160d1fb3def8133bdd0b66105e60ef93f80f82
Closes-Bug: #1481468