Updated openstack/openstack Project: openstack/nova  c7c0fe0847ec4b587d68b58ebc7cf0ad796c8ee5
libvirt: check min required qemu/libvirt versions on s390/s390x

The support for running KVM with libvirt on s390/s390x architectures
was documented in the original spec as requiring libvirt >= 1.0.4 and
qemu >= 1.4.0, however, no check was added to the init_host() method
in the libvirt driver. This could lead to users deploying Nova on
unsupported hypervisor platforms. KVM-on-Z developers have indicated
that in fact the min versions can be libvirt 1.2.13 and QEMU 2.3,
so this change adds such a check.

Blueprint: libvirt-kvm-systemz
Change-Id: I2851624a8cdaf353df3386e1e3e3cc0e35af41ec