Updated openstack/openstack Project: openstack/keystone  b121b9381a1d5f004abfd6fd072ce913bf887194
Correct enabled emulation query to request no attributes

The enabled emulation query was requesting that the cn attribute
be returned, but it didn't use the attribute value. This query
should have specified that no attributes be returned since this is
slightly more efficient as no extra data needs to be retrieved.

Closes-Bug: 1482687
Change-Id: I9f3cb80d02236498174e7962e3c1283386589ee2