Updated openstack/openstack Project: openstack/cinder  bc22b0111204124e154762ae6a17f9462dd42bd3
RBD: use user-configured value for chunk size

This patch allows a user to override the default chunk size in
cinder.conf.  When the driver is initialized, the default configuration
values in CONF are merged with the user-provided overrides into
self.configuration.  In most cases, self.configuration is used unless
the original default value is needed.  In this case, CONF was mistakenly
used when self.configuration was intended, which has the effect of
preventing the user from overriding the default chunk size by ignoring
the provided value.

Related-Bug: #1477706
Change-Id: I66a99950b7c4080de932be7ddac2f1595b9d59f4