Updated openstack/openstack Project: openstack/security-doc  1c58d01994ae6b0a7cbabd86981883d6876029ea
Update links that point to other documentation guides

In the Security Guide under the sub-category Management Utilities, there are two references that point
to the End User Guide. The first is for Command-line Clients Overview and the second is for
Download and Source the OpenStack RC file. The current links only redirect to the main page of the
OpenStack End-user Guide, not to the individual pages within the End-user Guide.

This bug fix is to update the links to point directly to the both reference sections in the User Guide
instead of the main page.

Change-Id: I44fa92af8497d3945a5af0b487ee209d430017b1
Closes-Bug: #1480471