Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  cc89ff771b29c892d002a0d457b09f8c94d6feff
Support anonymous credentials for Jenkins plugin info query

Allow setting user and password variables to 'None' to disable the auth
mechanism with python-jenkins.

Jenkins can be configured to allow read-only access to query the plugins
installed. This is useful in test mode to run with read-only access via
the special 'Anonymous' user group in Jenkins, so that the XML can be
generated based on the installed versions of plugins without any risk of
accidental update.

Change-Id: I98ed30dc248e704530429799ccde6347f074c59d