Updated openstack/openstack Project: openstack-infra/project-config  54f04e987f8754936302206534dba49c0fe542ba
Correct name of openstack-user-stories ACL file

To use the implicit ACL file lookup with jeepyb the ACL file name must
match that of the project name (with the added .config suffix). Move the
openstack-user-stories ACL file from user-stories.config to
openstack-user-stories.config so that it lines up and jeepyb can push
the ACLs.

Change-Id: I411e9f9386635c3285be588e194a0c12b8560e04