Updated openstack/openstack Project: openstack-infra/nodepool  1ef8c7fdaf37c4b6f337f7724610afa48a0ef885
Stop passing cloud config name to shade

In 150f266 we started passing the cloud config object into shade. This
causes a TypeError[1] because the keyword cloud_config argument
overrides the positional argument. We no longer need to pass in the
cloud_config name because the cloud_config object contains it and shade
is able retrieve the name from the object[2].

[1] http://paste.openstack.org/show/412441/
[2] http://git.openstack.org/cgit/openstack-infra/shade/tree/shade/__init__.py#n228

Change-Id: Ic10869245760d17ff285623d4558712ed1773239