Updated openstack/openstack Project: openstack/sahara  39f964e2e96c3a08580f0468089e1e11be660254
Check ACLs before adding access for Manila share

On investigation, Manila access calls are non-idempotent.
For this reason, it is vital that we assess the presence
or absence of pre-existing ACLs for any one instance before
calling share.allow(..., ip, ...) Otherwise subsequent
attempts to idempotently verify or add shares to a cluster
will fail.

Change-Id: Ib471876f896523856141434b343a743a0f357f84
Closes-bug: 1481849