Updated openstack/openstack Project: openstack/nova  9db64bb74c8f67ba3039a05f06faab3c872162ae
Ensure test_models_sync() works with new Alembic releases

test_models_sync() test is currently making assuptions about Alembic,
which might not be true in the upcoming releases: e.g. Column object
will not necessarily be bound to a Table instance.

Fix this by using the format described at:

http://alembic.readthedocs.org/en/rel_0_7/api.html#alembic.autogenerate.compare_metadata

Also, the code has been refactored a bit, so that we don't have to go
over a list of differences multiple times and it's not changed in-place
anymore (removing elements from the list while iterating over it is
error-prone on its own).

Closes-Bug: #1483287

Change-Id: Id0b8a9b7b11ee2cd249176cab403befdd9e06dbd