Updated openstack/openstack Project: openstack/nova  1dde750e019176f020a060a79febabd8b71e6391
libvirt: set caps on maximum live migration time

Currently Nova launches live migration and then just leaves it
to run, assuming it'll eventually finish. If the guest is
dirtying memory quicker than the network can transfer it, it
is entirely possible that a migration will never complete. In
such a case it is highly undesirable to leave it running
forever since it wastes valuable CPU and network resources
Rather than trying to come up with a new policy for aborting
migration in OpenStack, copy the existing logic that is
successfully battle tested by the oVirt project in VDSM.

This introduces two new host level configuration parameters
that cloud administrators can use to ensure migrations which
don't appear likely to complete will be aborted. First is
an overall cap on the total running time of migration. The
configured value is scaled by the number of GB of guest RAM,
since larger guests will obviously take proportionally longer
to migrate. The second is a timeout that is applied when Nova
detects that memory is being dirtied faster than it can be
transferred. It tracks a low watermark of data remaining and
if that low watermark doesn't decrease in the given time,
it assumes the VM is stuck and aborts migration

NB with the default values for the config parameters, the code
for detecting stuck migrations is only going to kick in for
guests which have more than 2 GB of RAM allocated, as for
smaller guests the overall time limit will abort migration
before this happens. This is reasonable as small guests are
less likely to get stuck during migration, as the network
will generally be able to keep up with dirtying data well
enough to get to a point where the final switchover can be
performed.

Related-bug: #1429220
DocImpact: two new libvirt configuration parameters in
           nova.conf allow the administrator to control
           length of migration before aborting it
Change-Id: I461affe6c85aaf2a6bf6e8749586bfbfe0ebc146