Updated openstack/openstack Project: openstack/keystone  0d3767fae9d2202fa5b4f63663df5f060d217910
Improve List Role Assignments Filters Performance

When listing role assignments, we have the option
to filter them by actor, target and role. The
current implementation uses the standard
filtering the V3.wrap_collection. Given the large
number of individual assignments, this is pretty
inefficient.

This patch allows the controller to pass the
filters into the driver call, so the driver
filter the results itself and as result, then the
list size is kept a minimum.

The restful testing that is already in place for
listing assignments will still act as an effective
test, since there is no change to external functionality
with this patch. However, a crucial byproduct of this
change is that follow-on patches can refactor a number
of manager methods related to effective assignments to
simply call this one rather than duplicate the algorithms
around the assignment backend. In order to ensure those
different usage patterns are tested, a separate series
of patches follow this to add manager-level testing of
this new assignment listing capability.

Co-Authored-By: Henry Nash <henryn@linux.vnet.ibm.com>
Co-Authored-By: Rodrigo Duarte <rodrigods@lsd.ufcg.edu.br>

Implements: bp list-role-assignments-performance

Change-Id: I20569a201f5cd4467a4891920255993c04dccda2