Updated openstack/openstack Project: openstack/horizon  dbf51538679a4937857c64ad225fa91d4a2921a2
JSCS Cleanup - style guide cleanup for Launch Instance wizard

Following John Papa's style guide
https://github.com/johnpapa/angular-styleguide,
this patch refactors the Angular code for Launch Instance wizard.

Note that the construction of the Launch Instance wizard precludes
applying JP's "controller as" rule to the top level wizard controller
itself; doing so breaks the *many* implicit parent accesses throughout
the wizard sub controllers (implicit parent scope accesses are not
possible using the "controller as" style). Fixing that is potentially
a flow-on tech-debt fix that is outside of the scope of this
patch - it'll be a bunch of work!

I also noticed a trivial bugfix while addressing this. It could be
broken out into a separate patch, but we have enough patches already ;)

Change-Id: Ie6a1fdca1da60259adbb65c36a9829b001ae752b
Partially-Implements: blueprint jscs-cleanup
Closes-Bug: 1481135