Updated openstack/openstack Project: openstack/ceilometer  b6af38032c38743258cf2550114e847fb259bf36
Correct thread handling in TranslationHook

TranslationHook was using a threading.local() to store an error
message however the threading.local() was being created in __init__
of the Hook. Hooks are not per request, so this wasn't really
working as planned.

Now, instead of using threading.local() at all, we just modify the
environ held by pecan (and later used by the
ParsableErrorMiddleware). environ is request-local.

This change feels a bit crufty because it is but because of the way
Exceptions are managed in the app there's not really any good way to do
it. We could consider changing the way Exceptions related to error
messages, but that would be a very large change.

Duplicated from similar change in aodh:
I463059df28f291cea0644b1a9908a907f146ba1f

Change-Id: Ie561ec7ee7ec8b6d4a535ba7b20569b5a1101878
Closes-Bug: #1481244