Updated openstack/openstack Project: openstack/nova  6ebd3ecf9a58dbe73c249c109208e42aea86f67f
libvirt: fix disk I/O QOS support with RBD

The disk I/O QOS settings were set in the libvirt_info
method of the Image class. While this is fine for most
subclasses, the Rbd subclass overrides this method and
so was loosing the QOS settings. Move the setting of
QOS parameters into a separate method so it can be
called in all places that need it. For added fun the
commit 86e6f34 which added QOS settings originally never
added any unit tests to cover its operation.

Co-authored: Daniel P. Berrange <berrange@redhat.com>
Fixes bug #1405367
Change-Id: Ibb3a4dff8996c29ef921be7c56648a442bbb89a2