Updated openstack/openstack Project: openstack-infra/system-config  d11fa5de0554c3582e914f17a40b7ef3113c9c09
Call register-zanata-projects on translate-dev

To allow us to register projects in Zanata automatically, pull in
jeepyb, and subscribe to the projects.yaml file to run it when it
changes.

This makes use of the wildfly user and home directory as a useful
place to copy and subscribe to the projects.yaml file.

Change-Id: I25bd20a38315819b79049dd947ada40029a92163
Depends-On: Id0229864b64ce39c70b559e1eb5d94822f1988ec
Depends-On: Id78a590ea07b886afc0c98e309ff6df45f2240bf