Updated openstack/openstack Project: openstack/openstack-manuals  bd1d7c4c62239fb10807a62bbc84fb8a55ad9d44
Include DiskFilter in the default list

Include DiskFilter in the default list DiskFilter is pretty critical
for proper operation of a deployment without shared storage, and is
likely to be something most new operators will want enabled. The
DiskFilter is now included in the default set in Nova. So adding it
to manual too.

Change-Id: I1ff200931299d390bc8e8d9759394e7e72e24f75
Closes-bug: #1483352