Updated openstack/openstack Project: openstack/keystone-specs  4a7b6ea57620d7754f70a0c465819326107bf104
Moves Dynamic Policy specs to Liberty dir

As per Keystone meeting today, we targeted the
Dynamic Policy specs to Liberty [1].

This change moves such specs to /liberty dir. In
addition, it renames the files to match
'Centralized Policies' instead of 'Dynamic
Policies', in order to better match the naming
defined during the review process and then the
title in the specs.

[1] http://eavesdrop.openstack.org/meetings/keystone/2015/keystone.2015-08-11-18.00.log.html

Change-Id: Ic6153e353cb610279e8849c6f99d910de7dfe8ed