Updated openstack/openstack Project: openstack/keystonemiddleware  9db204e5c2d5632ee8c3f316bae83ba0234fadab
Move common request processing to base class

Move the standard parts of token validation onto the base class. To make
this work in a way that is useful for both keystone and auth_token we
need to fetch and validate the tokens and then enforce what to do with
the request at a higher level.

Change-Id: I01ec6798b6c029c0013f4bcf73697c92a3063f4b