Updated openstack/openstack Project: openstack/nova  cbeecadae3c2dadbffcab3adb43a4de4ae604d5f
VMware: Do not untar OVA on the file system

The OVA image is a tarball which consists of OVF descriptor and VMDKs.
Currently the OVA is unpacked on the file system and then uploaded on
the datastore. This is inefficient and may cause n-cpu to run out of
space if large images are booted.
This patch fixes this by processing the OVA image as a stream -- we
subsequently read chunks of data from Glance and upload them to
vCenter.

Change-Id: I22f5897da0870a1d69fa153a0868000369159b37