Updated openstack/openstack Project: openstack/governance  62ca6deac471dbc3d5be4923142979a778e70ad4
diversity: Factor out policy from data gathering.

The diversity script had a single function that gathered a bunch of
stats from stackalytics and also determined whether the diversity
criteria was met.  Split the policy check out.  This is just a bit of
refactoring in preparation for adding some additional stats and policy
checks based on this code.

While we're at it, rename the script to 'teamstats', since this will
soon be looking at more than diversity.

Change-Id: I7993249b14cbb94a4700de10fc31bc0cd0394cf7
Signed-off-by: Russell Bryant <rbryant@redhat.com>