Updated openstack/openstack Project: openstack/governance  b163a41d79ae4627bd17cbfb2a0f09c892519847
teamstats: Add stats related to team size.

Update the teamstats script to be able to determine the number of
active committers and active reviewers.

An active committer is defined as someone with 6 commits over the last
6 months.  An active reviewer is defined as someone with 30 reviews
over the last 6 months.  These values are completely arbitrary.  They
don't matter yet since no tag is defined that depends on it.

Change-Id: I80c1ae2fb5e5b3fed20475777cfbc0fa7f70b9b3
Signed-off-by: Russell Bryant <rbryant@redhat.com>