Updated openstack/openstack Project: openstack/nova  607d83ba00d8a2c5c48b313ac63e8279a3da5810
cleanup NovaObjectDictCompat from numa object

cleanup subclassing on NovaObjectDictCompat and fix subsequent
tests and code associated with nova/objects/numa.py

Due to the exhaustive nature of changes, the cleanup is done
one object at a time.

This is the fifth patch of the series.
There shall be more patches to follow for other objects.

Related to blueprint liberty-objects

Change-Id: I7467a8640ea67f0a01751c2b3112ae56953978d1