Updated openstack/openstack Project: openstack/nova  737f22e5750a10614bf1b2e79d5108aaf2232ce3
Refactor unit test for InstanceGroup objects

The goal of this patch was to remove the use of the database in the unit
tests for the InstanceGroup objects. We now mock out all database
interactions in the object unit tests and verify that the calls to the
DB API are appropriate.

Change-Id: I98740029ed1115491438e93bfcb981b1f1bd7095