Updated openstack/openstack Project: openstack/neutron  a39c73ec6f19d9267d23192d2d03c65b3dc0a5c9
Revert "Revert "Add VIF_DELETED notification event to Nova""

This reverts commit 6575db592c92791a51540134192bc86465940283.

Depends-on: I998b6bb80cc0a81d665b61b8c4a424d7219c666f

DocImpact
If Neutron is upgraded to Liberty before the Nova API is,
the Nova API log will contain errors complaining that it doesn't
understand this new event. Nothing will be broken, but there will
be an error every time a port is deleted until Nova is upgraded.

Change-Id: I7aae44e62d2b1170bae31c3492148bfd516fb78b