Updated openstack/openstack Project: openstack/api-site  64246a22d43569e07a10828d7e0d69b73be9aabc
Always check all files with syntax check

The check-syntax only looks at files itself but if an entity file is
changed that is included by other XML files, we would need to check those as
well. Since this is not done currently and there are many entity files,
let's always check all files for syntax problems.

Change-Id: I34970df274e12a15a485bcd8fc3930651af64d4e