Updated openstack/openstack Project: openstack/swift  a2c0e6f68728f7837266f489098adf2cff78a500
Adding bandit for security static analysis testing in swift

This change adds a basic bandit config for Swift. It can be invoked
by running the tox environment for bandit;
    tox -e bandit

This is an initial step for using bandit with Swift
and it should be revisited to improve the testing as more is learned
about the specific needs of the Swift code base.As per now some tests
are excluded as they were used on purpose.

https://wiki.openstack.org/wiki/Security/Projects/Bandit

Implements: blueprint swift-bandit

Change-Id: I621be9a68ae9311f3a6eadd1636b05e646260cf2