Updated openstack/openstack Project: openstack/nova  5a494bdf0e5a2a45662c6434accbd0539d38f9b4
Add version counter to Service object

This implements blueprint service-version-number. It adds a monotonic
version number to the service object and data model. Existing non-versioned
service objects are said to be version 0, and anything running this patch
or later is version 1 (or later, when changed).

This doesn't affect anything specific yet, but can be used later to drive
policy (such as online data migrations, RPC version pins, etc) after this
is merged.

Change-Id: I12f3be255e37965270f9dd183a8a77b0774ce929